data-and-ai-jhb/oct/code/claude/expense-chatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
azure-ai-projects~=1.0.0b11
azure-identity~=1.20.0
ansible-core~=2.17.0
